CORE-8644 Move property and provider files into model package
