+++ de-lib/src/main/java/org/iplantc/de/apps/widgets/client/models/SelectionItemModelKeyProvider.java
@@ -1,4 +1,4 @@
-package org.iplantc.de.apps.widgets.client.view.editors;
+package org.iplantc.de.apps.widgets.client.models;
 
 import org.iplantc.de.client.models.apps.integration.SelectionItem;
