+++ de-lib/src/main/java/org/iplantc/de/apps/widgets/client/view/editors/arguments/IntegerSelectionEditor.java
@@ -1,6 +1,6 @@
 package org.iplantc.de.apps.widgets.client.view.editors.arguments;
 
-import org.iplantc.de.apps.widgets.client.view.editors.SelectionItemProperties;
+import org.iplantc.de.apps.widgets.client.models.SelectionItemProperties;
 import org.iplantc.de.apps.widgets.client.view.editors.style.AppTemplateWizardAppearance;
 import org.iplantc.de.apps.widgets.client.view.editors.widgets.AppWizardComboBox;
