+++ de-lib/src/main/java/org/iplantc/de/apps/widgets/client/view/editors/arguments/ReferenceAnnotationEditor.java
@@ -2,7 +2,7 @@ package org.iplantc.de.apps.widgets.client.view.editors.arguments;
 
 import static com.sencha.gxt.cell.core.client.form.ComboBoxCell.TriggerAction.ALL;
 
-import org.iplantc.de.apps.widgets.client.view.editors.ReferenceGenomeProperties;
+import org.iplantc.de.apps.widgets.client.models.ReferenceGenomeProperties;
 import org.iplantc.de.apps.widgets.client.view.editors.arguments.converters.ArgumentEditorConverter;
 import org.iplantc.de.apps.widgets.client.view.editors.arguments.converters.SplittableToReferenceGenomeConverter;
 import org.iplantc.de.apps.widgets.client.view.editors.style.AppTemplateWizardAppearance;
